aiohttp_tts_app/requirements.txt
aiodns==3.0.0
aiodogstatsd==0.16.0.post0
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-devtools==1.0.post0
aiohttp-jinja2==1.5
aioprometheus==22.5.0
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backports.zoneinfo==0.2.1
blinker==1.4
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
dataclasses-json==0.5.6
dateparser==1.0.0
devtools==0.8.0
dill==0.3.5.1
docopt==0.6.2
executing==0.8.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
frozenlist==1.3.0
greenlet==1.1.2
gruut==2.1.1
gruut-ipa==0.10.1
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.0
gruut-lang-it==2.0.0
gruut-lang-nl==2.0.0
gruut-lang-ru==2.0.0
gruut-lang-sv==2.0.0
gruut-lang-sw==2.0.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonlines==1.2.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
molotov==2.4
multidict==6.0.2
multiprocessing-on-dill==3.5.0a4
mypy-extensions==0.4.3
networkx==2.8.3
num2words==0.5.10
numpy==1.22.4
onnxruntime==1.10.0
orjson==3.7.2
phonemes2ids==1.0.0
Pillow==9.1.1
priority==2.0.0
prometheus-client==0.14.1
protobuf==3.19.4
psutil==5.8.0
pyaml==21.10.1
pycares==4.1.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyhumps==3.7.1
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-pidfile==3.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quantile-python==1.1
quart==0.17.0
Quart-CORS==0.5.0
quart-schema==0.11.1
regex==2022.6.2
shortuuid==1.0.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
swagger-ui-py==21.9.28
toml==0.10.2
tqdm==4.48.2
typing-inspect==0.7.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
watchgod==0.8.2
Werkzeug==2.0.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.8.0
